# performance tuning(mainly caching parameter's meta info)