--- a/src/SoapCore.BenchmarkOld/SoapCore.BenchmarkOld.csproj
+++ b/src/SoapCore.BenchmarkOld/SoapCore.BenchmarkOld.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="SoapCore" Version="0.9.7" />
+  </ItemGroup>
+	<ItemGroup>
+		<Compile Include="../SoapCore.Benchmark.Shared/*.cs"/>
+	</ItemGroup>
+	<ItemGroup>
+	  <PackageReference Include="BenchmarkDotNet" Version="0.10.13" />
+	  <PackageReference Include="Microsoft.AspNetCore" Version="2.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel.Core" Version="2.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.0.2" />
+	  <PackageReference Include="System.ServiceModel.Primitives" Version="4.4.1" />
+	</ItemGroup>
+
+</Project>