--- a/src/SoapCore/SoapEndpointMiddleware.cs
+++ b/src/SoapCore/SoapEndpointMiddleware.cs
@@ -151,12 +151,10 @@ namespace SoapCore
 			Message responseMessage;
 
 			//Reload the body to ensure we have the full message
-			using (var reader = new StreamReader(httpContext.Request.Body))
-			{
-				var body = await reader.ReadToEndAsync();
-				var requestData = Encoding.UTF8.GetBytes(body);
-				httpContext.Request.Body = new MemoryStream(requestData);
-			}
+			var mstm = new MemoryStream((int)httpContext.Request.ContentLength.GetValueOrDefault(1024));
+			await httpContext.Request.Body.CopyToAsync(mstm).ConfigureAwait(false);
+			mstm.Seek(0, SeekOrigin.Begin);
+			httpContext.Request.Body = mstm;
 
 			//Return metadata if no request
 			if (httpContext.Request.Body.Length == 0)
@@ -195,7 +193,8 @@ namespace SoapCore
 					// Get operation arguments from message
 					Dictionary<string, object> outArgs = new Dictionary<string, object>();
 					var arguments = GetRequestArguments(requestMessage, reader, operation, ref outArgs);
-					var allArgs = arguments.Concat(outArgs.Values).ToArray();
+					// avoid Concat() and ToArray() cost when no out args(this may be heavy operation)
+					var allArgs = outArgs.Count != 0 ? arguments.Concat(outArgs.Values).ToArray() : arguments;
 
 					// Invoke Operation method
 					var responseObject = operation.DispatchMethod.Invoke(serviceInstance, allArgs);
@@ -220,7 +219,7 @@ namespace SoapCore
 					}
 
 					// Create response message
-					var resultName = operation.DispatchMethod.ReturnParameter.GetCustomAttribute<MessageParameterAttribute>()?.Name ?? operation.Name + "Result";
+					var resultName = operation.ReturnName;
 					var bodyWriter = new ServiceBodyWriter(_serializer, operation.Contract.Namespace, operation.Name + "Response", resultName, responseObject, resultOutDictionary);
 					responseMessage = Message.CreateMessage(_messageEncoder.MessageVersion, null, bodyWriter);
 					responseMessage = new CustomMessage(responseMessage);
@@ -243,21 +242,20 @@ namespace SoapCore
 			return responseMessage;
 		}
 
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		private object[] GetRequestArguments(Message requestMessage, System.Xml.XmlDictionaryReader xmlReader, OperationDescription operation, ref Dictionary<string, object> outArgs)
 		{
-			var parameters = operation.DispatchMethod.GetParameters().Where(x => !x.IsOut && !x.ParameterType.IsByRef).ToArray();
-			var arguments = new List<object>();
+			var parameters = operation.NormalParameters;
+			// avoid reallocation
+			var arguments = new List<object>(parameters.Length + operation.OutParameters.Length);
 
 			// Find the element for the operation's data
 			xmlReader.ReadStartElement(operation.Name, operation.Contract.Namespace);
 
 			for (int i = 0; i < parameters.Length; i++)
 			{
-				var elementAttribute = parameters[i].GetCustomAttribute<XmlElementAttribute>();
-				var parameterName = !string.IsNullOrEmpty(elementAttribute?.ElementName)
-										? elementAttribute.ElementName
-										: parameters[i].GetCustomAttribute<MessageParameterAttribute>()?.Name ?? parameters[i].Name;
-				var parameterNs = elementAttribute?.Namespace ?? operation.Contract.Namespace;
+				var parameterName = parameters[i].Name;
+				var parameterNs = parameters[i].Namespace;
 
 				if (xmlReader.IsStartElement(parameterName, parameterNs))
 				{
@@ -265,9 +263,9 @@ namespace SoapCore
 
 					if (xmlReader.IsStartElement(parameterName, parameterNs))
 					{
-						var elementType = parameters[i].ParameterType.GetElementType();
-						if (elementType == null || parameters[i].ParameterType.IsArray)
-							elementType = parameters[i].ParameterType;
+						var elementType = parameters[i].Parameter.ParameterType.GetElementType();
+						if (elementType == null || parameters[i].Parameter.ParameterType.IsArray)
+							elementType = parameters[i].Parameter.ParameterType;
 
 						switch (_serializer)
 						{
@@ -294,16 +292,15 @@ namespace SoapCore
 				}
 			}
 
-			var outParams = operation.DispatchMethod.GetParameters().Where(x => x.IsOut || x.ParameterType.IsByRef).ToArray();
-			foreach (var parameterInfo in outParams)
+			foreach (var parameterInfo in operation.OutParameters)
 			{
-				if (parameterInfo.ParameterType.Name == "Guid&")
+				if (parameterInfo.Parameter.ParameterType.Name == "Guid&")
 					outArgs[parameterInfo.Name] = Guid.Empty;
-				else if (parameterInfo.ParameterType.Name == "String&" || parameterInfo.ParameterType.GetElementType().IsArray)
+				else if (parameterInfo.Parameter.ParameterType.Name == "String&" || parameterInfo.Parameter.ParameterType.GetElementType().IsArray)
 					outArgs[parameterInfo.Name] = null;
 				else
 				{
-					var type = parameterInfo.ParameterType.GetElementType();
+					var type = parameterInfo.Parameter.ParameterType.GetElementType();
 					outArgs[parameterInfo.Name] = Activator.CreateInstance(type);
 				}
 			}