following method can be heavy sometimes(found in my private investigates),
so I fixed to avoid these methods as possible as I could.
* `MethodInfo.GetParameters`(caching when startup)
* `ParameterInfo.GetCustomAttribute`(caching when startup)
* reloading stream with `StreamReader` and related(using `MemoryStream` directly)
* `IE<T>.Concat`(do not use when there are no 'out' parameter)
* `IE<T>.ToArray`(do not use when there are no 'out' parameter)

benchmark results in my development machine are as follows(10-15% improvement);

## before

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 8.1 (6.3.9600.0)
Intel Core i7-4770 CPU 3.40GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3312636 Hz, Resolution=301.8744 ns, Timer=TSC
.NET Core SDK=2.1.100
  [Host]     : .NET Core 2.0.5 (CoreCLR 4.6.26020.03, CoreFX 4.6.26018.01), 64bit RyuJIT
  Job-YACQPT : .NET Core 2.0.5 (CoreCLR 4.6.26020.03, CoreFX 4.6.26018.01), 64bit RyuJIT

TargetCount=5  

```
|    Method | LoopNum |      Mean |     Error |    StdDev |    Gen 0 | Allocated |
|---------- |-------- |----------:|----------:|----------:|---------:|----------:|
| EmptyTask |     100 |  2.349 ms | 0.6062 ms | 0.1575 ms | 242.1875 |   8.13 KB |
|      Echo |     100 | 10.843 ms | 0.2343 ms | 0.0609 ms | 921.8750 |   8.29 KB |

## after

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 8.1 (6.3.9600.0)
Intel Core i7-4770 CPU 3.40GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3312636 Hz, Resolution=301.8744 ns, Timer=TSC
.NET Core SDK=2.1.100
  [Host]     : .NET Core 2.0.5 (CoreCLR 4.6.26020.03, CoreFX 4.6.26018.01), 64bit RyuJIT
  Job-GNPZMU : .NET Core 2.0.5 (CoreCLR 4.6.26020.03, CoreFX 4.6.26018.01), 64bit RyuJIT

TargetCount=5  

```
|    Method | LoopNum |     Mean |     Error |    StdDev |    Gen 0 | Allocated |
|---------- |-------- |---------:|----------:|----------:|---------:|----------:|
| EmptyTask |     100 | 2.294 ms | 0.1480 ms | 0.0384 ms | 242.1875 |   8.13 KB |
|      Echo |     100 | 9.013 ms | 0.6711 ms | 0.1743 ms | 796.8750 |   8.29 KB |